+++ packages/riverpod/CHANGELOG.md
@@ -20,6 +20,8 @@ Non-breaking changes:
 - `AutoDisposeRef.maintainState` is deprecated. Use the new `AutoDisposeRef.keepAlive()` instead.
 - feat: Add support for `ref.invalidate(family)` to recompiute an entire family (#1517)
 - fixed a bug where `AsyncValue.whenData` did not preserve `AsyncValue.isLoading/isRefreshing`
+- fix: `StateProvider` and `StateNotifierProvider` no longer notify their listeners
+  on `ref.refresh` if the new result is identical to the old one.
 
 # 2.0.0-dev.9